Removing Merge and Split Activity Types
While important steps in entity management, we didn't have a clear direction for how to handle more complicated scenarios, e.g. partial splits, partial merges. We will defer this to a future round of specification development. The spec, as it is written, will will provision for full Splits and Merges using a combination of Deprecate and Create. For partial splits/merges (where the existing entity continues to exist), Update and Create can be used.