Published and Update date properties
Data Providers SHOULD include `published` on the the activity for when the activity was made available in the stream, and `updated` on the object for when the change to the entity actually occurred.

Removed `published` date from OrderedCollections, etc. because we're focusing on dates of Activities, when they were published and when the change actually occurred.